pyproject.toml
[project]
name = "contacts"
version = "0.1.0"
description = ""
authors = [{ name = "Osadchyi Yurii", email = "malicious1986@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "fastapi[standard] (>=0.121.1,<0.122.0)",
    "python-dotenv (>=1.0.0,<1.1.0)",
]

[tool.poetry]
packages = [{ include = "contacts", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
